fleshed out logic
A bit ugly, but the script now creates a grid based on a user-specified
order and density per order. This grid is searched and the best reward
is found. From there a new grid is made based on multipliers of the
best parameters and we search that grid for the new best parameters.